chore: Read delete file when `RowGroupReader::create_read_policy`
